added more unit-tests

+++ src/client/c/dptconvert.h
@@ -116,7 +116,7 @@ int16_t signed16FromPayload(const uint8_t *payload, int index);
 uint32_t unsigned32FromPayload(const uint8_t *payload, int index);
 int32_t signed32FromPayload(const uint8_t *payload, int index);
 double float16FromPayload(const uint8_t *payload, int index);
-double float32FromPayload(const uint8_t *payload, int index);
+float float32FromPayload(const uint8_t *payload, int index);
 int64_t signed64FromPayload(const uint8_t *payload, int index);
 uint8_t bcdFromPayload(const uint8_t *payload, int index);
